refactor: move PretrainedWeightsMixin to common + remove duplication
