+++ library/src/getitune/backend/lightning/models/classification/multilabel_models/mobilenet_v3.py
@@ -20,7 +20,7 @@ from getitune.backend.lightning.models.classification.losses.asymmetric_angular_
 from getitune.backend.lightning.models.classification.multilabel_models.base import LightningMultilabelClsModel
 from getitune.backend.lightning.models.classification.necks.gap import GlobalAveragePooling
 from getitune.backend.lightning.models.classification.utils.pretrained_urls import MOBILENETV3_PRETRAINED_URLS
-from getitune.backend.lightning.models.classification.utils.pretrained_weights import CheckpointWeightsLoader
+from getitune.backend.lightning.models.common.pretrained_weights import PretrainedWeightsMixin
 from getitune.backend.lightning.schedulers import LRSchedulerListCallable
 from getitune.data.entity.base import BatchLoss
 from getitune.data.entity.sample import PredictionBatch, SampleBatch
@@ -33,7 +33,7 @@ if TYPE_CHECKING:
     from getitune.metrics import MetricCallable
 
 
-class MobileNetV3MultilabelCls(CheckpointWeightsLoader, LightningMultilabelClsModel):
+class MobileNetV3MultilabelCls(PretrainedWeightsMixin, LightningMultilabelClsModel):
     """MobileNetV3 Model for multi-class classification task.
 
     Args:
@@ -50,6 +50,7 @@ class MobileNetV3MultilabelCls(CheckpointWeightsLoader, LightningMultilabelClsMo
         pretrained (bool, optional): Whether to use pretrained weights. Defaults to True.
     """
 
+    pretrained_weights_target = "backbone"
     pretrained_urls = MOBILENETV3_PRETRAINED_URLS
 
     def __init__(
