+++ library/src/getitune/backend/lightning/models/detection/base.py
@@ -23,7 +23,7 @@ from getitune.backend.lightning.models.base import (
     DefaultSchedulerCallable,
     LightningModel,
 )
-from getitune.backend.lightning.models.utils.pretrained_weights import PretrainedWeightsMixin
+from getitune.backend.lightning.models.common.pretrained_weights import PretrainedWeightsMixin
 from getitune.backend.lightning.models.utils.utils import InstanceData
 from getitune.backend.lightning.schedulers import LRSchedulerListCallable
 from getitune.backend.lightning.tools.explain.explain_algo import feature_vector_fn
